requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==4.1.0
attrs==24.2.0
behave==1.2.6
bleach==6.1.0
browserstack-local==1.2.8
browserstack-sdk==1.21.0
certifi==2024.8.30
charset-normalizer==3.3.2
dukpy==0.4.0
exceptiongroup==1.2.2
execnet==2.1.1
filelock==3.16.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
mutf8==1.0.6
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
parse_type==0.6.3
percy-appium-app==2.0.5
pillow==10.4.0
platformdirs==4.3.1
playwright==1.45.1
pluggy==1.5.0
psutil==6.0.0
pyee==11.1.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-SystemConfiguration==10.3.1
PyPAC==0.16.4
PySocks==1.7.1
pytest==7.4.2
pytest-base-url==2.1.0
pytest-bdd==7.2.0
pytest-playwright==0.5.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.24.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
tldextract==5.1.2
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
